refactor: 네비게이션 설정 공통 메소드 제거
코디네이터와 함께 사용 시 가독성이 오히려 저하되는 것 같아 제거했습니다.
