+++ Projects/TDPresentation/Sources/AppFlow/MainFlow/Home/HomeViewController.swift
@@ -1,4 +1,5 @@
 import TDDesign
+import TDCore
 import UIKit
 
 final class HomeViewController: BaseViewController<BaseView> {
@@ -37,9 +38,32 @@ final class HomeViewController: BaseViewController<BaseView> {
     }
     
     private func setupNavigationBar() {
-        setupNavigationBar(style: .home, navigationDelegate: coordinator!) { [weak self] in
+        // 좌측 네비게이션 바 버튼 설정 (캘린더 + 로고)
+        let calendarButton = UIButton(type: .custom)
+        calendarButton.setImage(TDImage.Calendar.top2Medium, for: .normal)
+        calendarButton.addAction(UIAction { [weak self] _ in
+            self?.coordinator?.didTapCalendarButton()
+        }, for: .touchUpInside)
+        
+        let toduckLogoImageView = UIImageView(image: TDImage.toduckLogo)
+        toduckLogoImageView.contentMode = .scaleAspectFit
+        
+        let leftBarButtonItems = [
+            UIBarButtonItem(customView: calendarButton),
+            UIBarButtonItem(customView: toduckLogoImageView)
+        ]
+        
+        navigationItem.leftBarButtonItems = leftBarButtonItems
+        
+        // 우측 네비게이션 바 버튼 설정 (알림)
+        let alarmButton = UIButton(type: .custom)
+        alarmButton.setImage(TDImage.Bell.topOffMedium, for: .normal)
+        alarmButton.addAction(UIAction { [weak self] _ in
             self?.coordinator?.didTapAlarmButton()
-        }
+            TDLogger.debug("홈 화면 네비게이션 우측 알람 버튼 클릭")
+        }, for: .touchUpInside)
+        
+        navigationItem.rightBarButtonItem = UIBarButtonItem(customView: alarmButton)
     }
     
     // MARK: - View Update
