+++ Projects/TDPresentation/Sources/AppFlow/MainFlow/MyPage/MyPageView/MyPageViewController.swift
@@ -1,17 +1,44 @@
 import UIKit
+import TDDesign
 import TDCore
 
 final class MyPageViewController: BaseViewController<MyPageView> {
     weak var coordinator: MyPageCoordinator?
     
-    override func addView() {
-        setupNavigationBar(style: .mypage, navigationDelegate: coordinator!) {
-            TDLogger.debug("MyPageViewController - setupNavigationBar")
-        }
-    }
-    
     override func layout() {
         self.view.backgroundColor = .white
+    }
+    
+    override func configure() {
+        setupNavigationBar()
+    }
+    
+    private func setupNavigationBar() {
+        // 좌측 네비게이션 바 버튼 설정 (캘린더 + 로고)
+        let calendarButton = UIButton(type: .custom)
+        calendarButton.setImage(TDImage.Calendar.top2Medium, for: .normal)
+        calendarButton.addAction(UIAction { [weak self] _ in
+            self?.coordinator?.didTapCalendarButton()
+        }, for: .touchUpInside)
+        
+        let toduckLogoImageView = UIImageView(image: TDImage.toduckLogo)
+        toduckLogoImageView.contentMode = .scaleAspectFit
+        
+        let leftBarButtonItems = [
+            UIBarButtonItem(customView: calendarButton),
+            UIBarButtonItem(customView: toduckLogoImageView)
+        ]
+        
+        navigationItem.leftBarButtonItems = leftBarButtonItems
+        
+        // 우측 네비게이션 바 버튼 설정 (알림 버튼)
+        let alarmButton = UIButton(type: .custom)
+        alarmButton.setImage(TDImage.Bell.topOffMedium, for: .normal)
+        alarmButton.addAction(UIAction { _ in
+            TDLogger.debug("MyPageViewController - 알람 버튼 클릭")
+        }, for: .touchUpInside)
+        
+        navigationItem.rightBarButtonItem = UIBarButtonItem(customView: alarmButton)
     }
     
     override func touchesEnded(_ touches: Set<UITouch>, with event: UIEvent?) {
